punto de venta testeada

+++ src/sectorDeSaldos/IControlSaldo.java
@@ -2,8 +2,8 @@ package sectorDeSaldos;
 
 public interface IControlSaldo {
 
-	public Integer saldo(String numCelular);
-	public void descontar(String numCelular, Integer cantADescontar);
-	public void cargarSaldo(String numCelular, Integer saldoACargar);
+	public Double saldo(String numCelular);
+	public void descontar(String numCelular, Double cantADescontar);
+	public void cargarSaldo(String numCelular, Double saldoACargar);
 	
 }
