+++ src/sectorDeSaldos/sectorDeSaldo.java
@@ -6,46 +6,46 @@ import java.util.Map;
 public class sectorDeSaldo implements ISectorSaldo {
 
 	//atributos 
-	Map<String, Integer> creditoCelulares;
+	Map<String, Double> creditoCelulares;
 	
 	
 	//constructor 
-	public sectorDeSaldo(Map<String, Integer>celularesYCreditos) {
+	public sectorDeSaldo(Map<String, Double>celularesYCreditos) {
 		this.setCreditoCelulares(celularesYCreditos);
 	}
 	
 	//metodos
-	public Map<String, Integer> getSaldos(){
+	public Map<String, Double> getSaldos(){
 		return this.getCreditoCelulares();
 	}
 
-	public Map<String, Integer> getCreditoCelulares() {
+	public Map<String, Double> getCreditoCelulares() {
 		return creditoCelulares;
 	}
 
-	public void setCreditoCelulares(Map<String, Integer> creditoCelulares) {
+	public void setCreditoCelulares(Map<String, Double> creditoCelulares) {
 		this.creditoCelulares = creditoCelulares;
 	}
 
-	public Integer saldo(String numCelular) {
+	public Double saldo(String numCelular) {
 		return this.getCreditoCelulares().get(numCelular);
 	}
 	
-	public void descontar(String numCelular, Integer cantADescontar) {
-		Integer valorCambiado = (this.getCreditoCelulares().get(numCelular)) - cantADescontar;
+	public void descontar(String numCelular, Double cantADescontar) {
+		Double valorCambiado = (this.getCreditoCelulares().get(numCelular)) - cantADescontar;
 		this.getCreditoCelulares().remove(numCelular);
 		this.getCreditoCelulares().put(numCelular, valorCambiado);
 	}
 	
-	public void cargarSaldo(String numCelular, Integer saldoACargar) {
-		Integer valorCambiado = (this.getCreditoCelulares().get(numCelular)) + saldoACargar;
+	public void cargarSaldo(String numCelular, Double saldoACargar) {
+		Double valorCambiado = (this.getCreditoCelulares().get(numCelular)) + saldoACargar;
 		this.getCreditoCelulares().remove(numCelular);
 		this.getCreditoCelulares().put(numCelular, valorCambiado);
 	}
 
-	public Boolean creditoSuficiente(Integer cantDeHoras, String numCelular) {
-		Integer costoTotal = cantDeHoras * 40;
-		Integer valor = this.getCreditoCelulares().get(numCelular);
+	public Boolean creditoSuficiente(Double cantDeHoras, String numCelular) {
+		Double costoTotal = cantDeHoras * 40;
+		Double valor = this.getCreditoCelulares().get(numCelular);
 		return costoTotal < valor;
 	}
 }
